backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymongo==4.4.0
pydantic>=2.0.0
requests==2.31.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-json-logger==2.0.7
annotated-types>=0.4.0
